Refactor Common and Dummy gateway request classes
