+++ src/Omnipay/AuthorizeNet/AIMResponse.php
@@ -11,7 +11,7 @@
 
 namespace Omnipay\AuthorizeNet;
 
-use Omnipay\Common\AbstractResponse;
+use Omnipay\Common\Message\AbstractResponse;
 use Omnipay\Exception;
 use Omnipay\Common\Exception\InvalidResponseException;
